Inet: Split EndPoint into per-implementation files
#### Problem

Part of #7715 _Virtualize System and Inet interfaces_

#### Change overview

Move each `UDPEndPoint` and `TCPEndPoint` implementation into a
separate file. (This follows the pattern used by `System::Layer`.)

#### Testing

CI; no change to functionality.
