+++ src/inet/BUILD.gn
@@ -49,6 +49,11 @@ buildconfig_header("inet_buildconfig") {
     defines +=
         [ "INET_PLATFORM_CONFIG_INCLUDE=${chip_inet_platform_config_include}" ]
   }
+
+  defines += [
+    "INET_TCP_END_POINT_IMPL_CONFIG_FILE=<inet/TCPEndPointImpl${chip_system_config_inet}.h>",
+    "INET_UDP_END_POINT_IMPL_CONFIG_FILE=<inet/UDPEndPointImpl${chip_system_config_inet}.h>",
+  ]
 }
 
 source_set("inet_config_header") {
@@ -105,6 +110,9 @@ static_library("inet") {
     sources += [
       "TCPEndPoint.cpp",
       "TCPEndPoint.h",
+      "TCPEndPointImpl${chip_system_config_inet}.cpp",
+      "TCPEndPointImpl${chip_system_config_inet}.h",
+      "TCPEndPointImpl.h",
     ]
   }
 
@@ -112,6 +120,9 @@ static_library("inet") {
     sources += [
       "UDPEndPoint.cpp",
       "UDPEndPoint.h",
+      "UDPEndPointImpl${chip_system_config_inet}.cpp",
+      "UDPEndPointImpl${chip_system_config_inet}.h",
+      "UDPEndPointImpl.h",
     ]
   }
