+++ src/inet/TCPEndPoint.h
@@ -36,17 +36,6 @@
 
 #include <utility>
 
-#if CHIP_SYSTEM_CONFIG_USE_LWIP
-#include <inet/EndPointStateLwIP.h>
-#endif // CHIP_SYSTEM_CONFIG_USE_LWIP
-#if CHIP_SYSTEM_CONFIG_USE_SOCKETS
-#include <inet/EndPointStateSockets.h>
-#endif // CHIP_SYSTEM_CONFIG_USE_SOCKETS
-
-#if CHIP_SYSTEM_CONFIG_USE_DISPATCH
-#include <dispatch/dispatch.h>
-#endif
-
 namespace chip {
 
 namespace Transport {
@@ -698,136 +687,6 @@ protected:
     virtual void HandleConnectCompleteImpl()                                                                   = 0;
     virtual void DoCloseImpl(CHIP_ERROR err, State oldState)                                                   = 0;
 };
-
-#if CHIP_SYSTEM_CONFIG_USE_LWIP
-
-class TCPEndPointImplLwIP : public TCPEndPoint, public EndPointStateLwIP
-{
-public:
-    TCPEndPointImplLwIP(EndPointManager<TCPEndPoint> & endPointManager) : TCPEndPoint(endPointManager), mUnackedLength(0) {}
-
-    // TCPEndPoint overrides.
-    CHIP_ERROR GetPeerInfo(IPAddress * retAddr, uint16_t * retPort) const override;
-    CHIP_ERROR GetLocalInfo(IPAddress * retAddr, uint16_t * retPort) const override;
-    CHIP_ERROR GetInterfaceId(InterfaceId * retInterface) override;
-    CHIP_ERROR EnableNoDelay() override;
-    CHIP_ERROR EnableKeepAlive(uint16_t interval, uint16_t timeoutCount) override;
-    CHIP_ERROR DisableKeepAlive() override;
-    CHIP_ERROR AckReceive(uint16_t len) override;
-#if INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-    void TCPUserTimeoutHandler() override;
-#endif // INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-
-private:
-    // TCPEndPoint overrides.
-    CHIP_ERROR BindImpl(IPAddressType addrType, const IPAddress & addr, uint16_t port, bool reuseAddr) override;
-    CHIP_ERROR ListenImpl(uint16_t backlog) override;
-    CHIP_ERROR ConnectImpl(const IPAddress & addr, uint16_t port, InterfaceId intfId) override;
-    CHIP_ERROR SendQueuedImpl(bool queueWasEmpty) override;
-    CHIP_ERROR SetUserTimeoutImpl(uint32_t userTimeoutMillis) override;
-    CHIP_ERROR DriveSendingImpl() override;
-    void HandleConnectCompleteImpl() override;
-    void DoCloseImpl(CHIP_ERROR err, State oldState) override;
-
-    struct BufferOffset
-    {
-        BufferOffset(System::PacketBufferHandle && aBuffer) : buffer(std::move(aBuffer)), offset(0) {}
-        BufferOffset(BufferOffset && aOther)
-        {
-            buffer = std::move(aOther.buffer);
-            offset = aOther.offset;
-        }
-        chip::System::PacketBufferHandle buffer;
-        uint16_t offset;
-    };
-
-    uint16_t mUnackedLength; // Amount sent but awaiting ACK. Used as a form of reference count
-                             // to hang-on to backing packet buffers until they are no longer needed.
-
-    uint16_t RemainingToSend();
-    BufferOffset FindStartOfUnsent();
-    CHIP_ERROR GetPCB(IPAddressType addrType);
-    void HandleDataSent(uint16_t len);
-    void HandleDataReceived(chip::System::PacketBufferHandle && buf);
-    void HandleIncomingConnection(TCPEndPoint * pcb);
-    void HandleError(CHIP_ERROR err);
-
-    static err_t LwIPHandleConnectComplete(void * arg, struct tcp_pcb * tpcb, err_t lwipErr);
-    static err_t LwIPHandleIncomingConnection(void * arg, struct tcp_pcb * tcpConPCB, err_t lwipErr);
-    static err_t LwIPHandleDataReceived(void * arg, struct tcp_pcb * tpcb, struct pbuf * p, err_t err);
-    static err_t LwIPHandleDataSent(void * arg, struct tcp_pcb * tpcb, u16_t len);
-    static void LwIPHandleError(void * arg, err_t err);
-};
-
-using TCPEndPointImpl = TCPEndPointImplLwIP;
-
-#endif // CHIP_SYSTEM_CONFIG_USE_LWIP
-
-#if CHIP_SYSTEM_CONFIG_USE_SOCKETS
-
-class TCPEndPointImplSockets : public TCPEndPoint, public EndPointStateSockets
-{
-public:
-    TCPEndPointImplSockets(EndPointManager<TCPEndPoint> & endPointManager) :
-        TCPEndPoint(endPointManager)
-#if INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-        ,
-        mBytesWrittenSinceLastProbe(0), mLastTCPKernelSendQueueLen(0)
-#endif // INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-    {}
-
-    // TCPEndPoint overrides.
-    CHIP_ERROR GetPeerInfo(IPAddress * retAddr, uint16_t * retPort) const override;
-    CHIP_ERROR GetLocalInfo(IPAddress * retAddr, uint16_t * retPort) const override;
-    CHIP_ERROR GetInterfaceId(InterfaceId * retInterface) override;
-    CHIP_ERROR EnableNoDelay() override;
-    CHIP_ERROR EnableKeepAlive(uint16_t interval, uint16_t timeoutCount) override;
-    CHIP_ERROR DisableKeepAlive() override;
-    CHIP_ERROR AckReceive(uint16_t len) override;
-#if INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-    void TCPUserTimeoutHandler() override;
-#endif // INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-
-private:
-    // TCPEndPoint overrides.
-    CHIP_ERROR BindImpl(IPAddressType addrType, const IPAddress & addr, uint16_t port, bool reuseAddr) override;
-    CHIP_ERROR ListenImpl(uint16_t backlog) override;
-    CHIP_ERROR ConnectImpl(const IPAddress & addr, uint16_t port, InterfaceId intfId) override;
-    CHIP_ERROR SendQueuedImpl(bool queueWasEmpty) override;
-    CHIP_ERROR SetUserTimeoutImpl(uint32_t userTimeoutMillis) override;
-    CHIP_ERROR DriveSendingImpl() override;
-    void HandleConnectCompleteImpl() override;
-    void DoCloseImpl(CHIP_ERROR err, State oldState) override;
-
-    CHIP_ERROR GetSocketInfo(int getname(int, sockaddr *, socklen_t *), IPAddress * retAddr, uint16_t * retPort) const;
-    CHIP_ERROR GetSocket(IPAddressType addrType);
-    void HandlePendingIO(System::SocketEvents events);
-    void ReceiveData();
-    void HandleIncomingConnection();
-    CHIP_ERROR BindSrcAddrFromIntf(IPAddressType addrType, InterfaceId intfId);
-    static void HandlePendingIO(System::SocketEvents events, intptr_t data);
-
-#if CHIP_SYSTEM_CONFIG_USE_DISPATCH
-    dispatch_source_t mReadableSource  = nullptr;
-    dispatch_source_t mWriteableSource = nullptr;
-#endif // CHIP_SYSTEM_CONFIG_USE_DISPATCH
-
-#if INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-    /// This counts the number of bytes written on the TCP socket since thelast probe into the TCP outqueue was made.
-    uint32_t mBytesWrittenSinceLastProbe;
-
-    /// This is the measured size(in bytes) of the kernel TCP send queue at the end of the last user timeout window.
-    uint32_t mLastTCPKernelSendQueueLen;
-
-    CHIP_ERROR CheckConnectionProgress(bool & IsProgressing);
-#endif // INET_CONFIG_OVERRIDE_SYSTEM_TCP_USER_TIMEOUT
-};
-
-using TCPEndPointImpl = TCPEndPointImplSockets;
-
-#endif // CHIP_SYSTEM_CONFIG_USE_SOCKETS
-
-using TCPEndPointManagerImpl = EndPointManagerImplPool<TCPEndPointImpl, INET_CONFIG_NUM_TCP_ENDPOINTS>;
 
 template <>
 struct EndPointProperties<TCPEndPoint>
